[exporter/datadogexporter] Move all packages but config to internal folder
