+++ exporter/datadogexporter/internal/metadata/azure/azure.go
