+++ exporter/datadogexporter/internal/metadata/system/host_windows.go
