+++ exporter/datadogexporter/internal/metadata/valid/valid.go
