+++ exporter/datadogexporter/internal/metadata/valid/valid_test.go
