+++ exporter/datadogexporter/internal/utils/http_test.go
