+++ exporter/datadogexporter/internal/utils/trace_helpers.go
