+++ exporter/datadogexporter/internal/utils/trace_helpers_test.go
